dom/security/test/csp/test_empty_directive.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1439425
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1439425</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1439425">Mozilla Bug 1439425</a>
<p id="display"></p>

<iframe id="cspframe"></iframe>

<pre id="test">

<script class="testbody" type="text/javascript">
let consoleCount = 0;

function cleanup() {
  SpecialPowers.postConsoleSentinel();
}

function finish() {
  SimpleTest.finish();
}

SpecialPowers.registerConsoleListener(function ConsoleMsgListener(aMsg) {
  if (aMsg.message == "SENTINEL") {
    is(consoleCount, 0);
    SimpleTest.executeSoon(finish);
  } else if (aMsg.message.includes("Content Security Policy")) {
    ++consoleCount;
    ok(false, "Must never see a console warning here");
  }
});

// set up and start testing
SimpleTest.waitForExplicitFinish();
let frame = document.getElementById('cspframe');
frame.onload = () => {
  SimpleTest.executeSoon(cleanup);
};
frame.src = 'file_empty_directive.html';
</script>
</pre>
</body>
</html>
